WIP: implement react context for storing rowinfo and selectedrows
